Give cars a driver with a FOV.
Each car is initialised with a driver, which is updated during scenario.step. The driver has a Shape that represents its field of view, and will have access to an up to date list of all other fixtures in that field of view. (Currently also wheels, I should exclude those.) The drivers share common knowledge through the scenario, and common knowledge could be info such as the mouse position.